myweb/requirements.txt
asgiref==3.8.1
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
Django==5.0.3
django-extensions==3.2.3
django-filter==24.1
djangorestframework==3.15.1
djangorestframework-jsonapi==6.1.0
environs==11.0.0
executing==1.2.0
fonttools==4.40.0
inflection==0.5.1
ipykernel==6.23.1
ipython==8.14.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
marshmallow==3.21.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mysqlclient==2.1.1
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
pandas==2.0.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.5.3
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pyzmq==25.1.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
threadpoolctl==3.1.0
tornado==6.3.2
traitlets==5.9.0
tzdata==2024.1
wcwidth==0.2.6
